% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ena_archivos}
\alias{ena_archivos}
\title{Encontrar archivos de datos ENA}
\usage{
ena_archivos(ruta, recursivo = TRUE)
}
\arguments{
\item{ruta}{Archivo o directorio de una descarga ENA.}

\item{recursivo}{Buscar en subdirectorios.}
}
\value{
Un vector nombrado de rutas por extension.
}
\description{
Encontrar archivos de datos ENA
}
