% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis.R
\name{ena_combinar}
\alias{ena_combinar}
\title{Combinar bases ENA de manera segura}
\usage{
ena_combinar(datos, id_origen = "origen")
}
\arguments{
\item{datos}{Lista de \code{data.frame} o varios \code{data.frame}.}

\item{id_origen}{Nombre de la columna que identifica la fuente.}
}
\value{
Un \code{data.frame} combinado.
}
\description{
Agrega una columna de origen antes de apilar tablas con columnas distintas.
}
