% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis.R
\name{ena_estimar_media}
\alias{ena_estimar_media}
\title{Estimar una media ponderada}
\usage{
ena_estimar_media(datos, variable, pesos, por = NULL)
}
\arguments{
\item{datos}{Un \code{data.frame} o un objeto de \code{\link[=ena_diseno]{ena_diseno()}}.}

\item{variable}{Variable numerica.}

\item{pesos}{Variable de ponderacion si \code{datos} no es \code{ena_diseno}.}

\item{por}{Vector opcional de dominios.}
}
\value{
Un \code{data.frame} con tamano muestral, suma de pesos y media.
}
\description{
Calcula la media descriptiva ponderada, opcionalmente por dominios. El
resultado no incluye error estandar porque este depende del diseno muestral
especifico de cada anio ENA.
}
