% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enaR-package.R
\docType{package}
\name{enaPeruR-package}
\alias{enaPeruR}
\alias{enaPeruR-package}
\title{enaPeruR: microdatos de la Encuesta Nacional Agropecuaria}
\description{
Infraestructura reproducible para trabajar con los microdatos ENA del INEI.
La cobertura incluida es 2014--2025. Antes de unir periodos, consulte
\code{\link[=ena_marco]{ena_marco()}} y documente el tratamiento de la ruptura metodologica de 2023.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://proyectos.inei.gob.pe/microdatos/}
  \item Report bugs at \url{https://proyectos.inei.gob.pe/microdatos/}
}

}
\author{
\strong{Maintainer}: Equipo enaR \email{maintainer@example.com}

Authors:
\itemize{
  \item Equipo enaR \email{maintainer@example.com}
}

}
\keyword{internal}
