% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis.R
\name{ena_estimar_proporcion}
\alias{ena_estimar_proporcion}
\title{Estimar proporciones ponderadas}
\usage{
ena_estimar_proporcion(datos, variable, pesos, por = NULL, porcentaje = FALSE)
}
\arguments{
\item{datos}{Un \code{data.frame} o un objeto de \code{\link[=ena_diseno]{ena_diseno()}}.}

\item{variable}{Variable numerica.}

\item{pesos}{Variable de ponderacion si \code{datos} no es \code{ena_diseno}.}

\item{por}{Vector opcional de dominios.}

\item{porcentaje}{Si es \code{TRUE}, expresa la proporcion en porcentaje.}
}
\value{
Un \code{data.frame} ordenado por dominio y categoria.
}
\description{
Calcula frecuencias muestrales, poblacion expandida y proporciones
ponderadas para una variable categorica.
}
