% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enaR-package.R
\docType{package}
\name{enaPeruR-package}
\alias{enaPeruR}
\alias{enaPeruR-package}
\title{enaPeruR: microdatos de la Encuesta Nacional Agropecuaria}
\description{
Infraestructura reproducible para descargar, leer, validar, apilar y tabular
microdatos ENA del INEI. La cobertura incluida es 2014--2025. El paquete
facilita la preparacion de bases multianuales, pero no armoniza cuestionarios,
codigos, factores de expansion o cambios de cobertura. Antes de comparar
periodos, consulte \code{\link[=ena_catalogo]{ena_catalogo()}} y \code{\link[=ena_marco]{ena_marco()}} y documente el tratamiento
de la ruptura metodologica de 2023.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://proyectos.inei.gob.pe/microdatos/}
  \item \url{https://paulesantos.github.io/enaPeruR/}
  \item Report bugs at \url{https://proyectos.inei.gob.pe/microdatos/}
}

}
\author{
\strong{Maintainer}: Paul Efren Santos Andrade \email{paulefrens@gmail.com} (\href{https://orcid.org/0000-0002-6635-0375}{ORCID})

Authors:
\itemize{
  \item Paul Efren Santos Andrade \email{paulefrens@gmail.com} (\href{https://orcid.org/0000-0002-6635-0375}{ORCID})
}

}
\keyword{internal}
